requirements.txt
streamlit
groq
youtube-transcript-api==0.6.3
python-dotenv
yt-dlp
